Hell yeah!!!!!
Removed lib of JgrapX from Jgraph module and added sources to the same module.

Added default constructor to mxGraphComponent. It's the same as a constructor with mxGraph parameter, instead of it a new graph is created.

Added mxGraphComponent to VisualEditor TopComponent, and made a minor fixes to show Ihor's example.